fix: 필드 타입과 상수 타입 불일치 수정

+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/dto/ReviewStats.java
@@ -7,7 +7,7 @@ public record ReviewStats(
 ) {
 
     private static final double DEFAULT_REVIEW_AVERAGE = 0.0;
-    private static final int DEFALUT_REVIEW_COUNT = 0;
+    private static final long DEFALUT_REVIEW_COUNT = 0;
 
     public static ReviewStats defaultOf(Long mentoringId) {
         return new ReviewStats(mentoringId, DEFAULT_REVIEW_AVERAGE, DEFALUT_REVIEW_COUNT);
